% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_calcs.R
\name{calc_entropy}
\alias{calc_entropy}
\title{Calculate Entropy}
\usage{
calc_entropy(counts)
}
\arguments{
\item{counts}{A matrix of count data}
}
\value{
A vector of entropies
}
\description{
This is a description
}
\examples{

mat <- matrix(1:12, ncol = 3)
calc_entropy(mat)

}
