% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_calcs.R
\name{select_entropy}
\alias{select_entropy}
\title{Select Genes By Entropy}
\usage{
select_entropy(counts, threshold)
}
\arguments{
\item{counts}{counts matrix}

\item{threshold}{Threshold}
}
\value{
Vector of truth values
}
\description{
This is a. escription
}
\examples{

mat <- matrix(1:12, ncol = 3)
select_entropy(mat, 1.3)

}
